+++ database/migrations/2021_06_26_075924_add_collections_field.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddCollectionsField extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('challenges', function (Blueprint $table) {
+            $table->integer('collectionable')->default(0)->after('challenge_required');
+            $table->unsignedBigInteger('collection')->nullable()->after('collectionable');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+
+    }
+}