+++ resources/views/collections/index.blade.php
@@ -18,6 +18,14 @@
         <span>{{ __('menu.collections') }}</span>
     </a>
 </li>
+<span class="left-auto is-flex">
+    <a @click="$refs.collections.$data.isPrefsModalActive = true">
+        <span class="icon is-small">
+            <i class="mx-2 fas fa-cog"></i>
+        </span>
+        <span class="is-hidden-mobile">{{ __('menu.config') }}</span>
+    </a>
+</span>
 @include('layouts.breadcrumbend')
 @endsection
 
@@ -25,5 +33,5 @@
 @endsection
 
 @section('content')
-    <show-collections code="{{ $class->code }}" :collections="{{ json_encode($collections) }}"></show-collections>
+    <show-collections ref="collections" :settings="{{ json_encode($settings) }}" code="{{ $class->code }}" :collections="{{ json_encode($collections) }}"></show-collections>
 @endsection